requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.3.2
pymongo==4.6.1
openai==1.12.0
python-dotenv==1.0.1
pydantic==1.10.13
httpx==0.24.1
mangum==0.17.0
python-multipart==0.0.9
typing-extensions==4.7.1
starlette==0.27.0
anyio==3.7.1
click==8.1.7
h11==0.14.0
sniffio==1.3.0
tqdm==4.65.0
certifi==2023.7.22
httpcore==0.17.3
idna==3.4
colorama==0.4.6
annotated-types==0.7.0
sounddevice==0.4.6
numpy==1.26.4
SpeechRecognition==3.10.0
python-docx
langchain-openai
qdrant-client